requirements.txt
numpy
matplotlib
opencv-python
Pillow
scikit-learn
streamlit
tensorflow